+++ src/model/entity/Customer.java
@@ -2,7 +2,7 @@ package model.entity;
 
 public class Customer extends Person {
 
-    double id;
+    long id;
 
     public Customer() {
     }
@@ -11,11 +11,11 @@ public class Customer extends Person {
         super(cpf);
     }
 
-    public Customer(double id) {
+    public Customer(long id) {
         this.id = id;
     }
 
-    public Customer(String cpf, double id) {
+    public Customer(String cpf, long id) {
         super(cpf);
         this.id = id;
     }
@@ -24,7 +24,7 @@ public class Customer extends Person {
         return id;
     }
 
-    public void setId(double id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
